requirements.txt
# Core dependencies
streamlit>=1.29.0
langgraph>=0.5.0
langchain>=0.3.26
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.3.0

# Database
chromadb>=0.4.18
sqlalchemy>=2.0.23

# ML/AI models
sentence-transformers>=2.2.2
transformers>=4.36.2
torch>=2.1.2
FlagEmbedding>=1.2.10

# HTTP Client
httpx>=0.25.2

# Utilities
python-dotenv>=1.0.0
pandas>=2.1.4
numpy>=1.24.4
requests>=2.31.0

# Development
pytest>=7.4.3
black>=23.12.0
flake8>=6.1.0 